Refactor: move methods to turn_sequence.cpp
